Crate prio

libprio-rs

Implementation of the Prio private data aggregation protocol.

Prio3 is available in the vdaf module as part of an implementation of Verifiable Distributed Aggregation Functions, along with an experimental implementation of Poplar1.

Modules

codec

Support for encoding and decoding messages to or from the TLS wire encoding, as specified in RFC 8446, Section 3.

dp

Differential privacy (DP) primitives.

field

Finite field arithmetic.

flp

Implementation of the generic Fully Linear Proof (FLP) system specified in [draft-irtf-cfrg-vdaf-07]. This is the main building block of Prio3.

idpf

This module implements the incremental distributed point function (IDPF) described in [draft-irtf-cfrg-vdaf-07].

topology

Implementations of some aggregator communication topologies specified in VDAF.

vdaf

Verifiable Distributed Aggregation Functions (VDAFs) as described in [draft-irtf-cfrg-vdaf-07].