Trait prio::codec::ParameterizedEncode

pub trait ParameterizedEncode<P> {
    fn encode_with_param(&self, encoding_parameter: &P, bytes: &mut Vec<u8>);

    fn get_encoded_with_param(&self, encoding_parameter: &P) -> Vec<u8> { ... }
    fn encoded_len_with_param(&self, _encoding_parameter: &P) -> Option<usize> { ... }
}

Describes how to encode objects into a byte sequence.

Required Methods

fn encode_with_param(&self, encoding_parameter: &P, bytes: &mut Vec<u8>)

Append the encoded form of this object to the end of bytes, growing the vector as needed. encoding_parameter provides details of the wire encoding, used to control how the value is encoded.

Provided Methods

fn get_encoded_with_param(&self, encoding_parameter: &P) -> Vec<u8>

Convenience method to encode a value into a new Vec<u8>.

fn encoded_len_with_param(&self, _encoding_parameter: &P) -> Option<usize>

Returns an optional hint indicating how many bytes will be required to encode this value, or None by default.