Trait prio::dp::DifferentialPrivacyStrategy

pub trait DifferentialPrivacyStrategy {
    type Budget: DifferentialPrivacyBudget;
    type Distribution: DifferentialPrivacyDistribution;
    type Sensitivity;

    // Required methods
    fn from_budget(b: Self::Budget) -> Self;
    fn create_distribution(
        &self,
        s: Self::Sensitivity
    ) -> Result<Self::Distribution, DpError>;
}

Strategy to make aggregate results differentially private, e.g. by adding noise from a specific type of distribution instantiated with a given DP budget.

Required Associated Types

type Budget: DifferentialPrivacyBudget

The type of the DP budget, i.e. the variant of differential privacy that can be obtained by using this strategy.

type Distribution: DifferentialPrivacyDistribution

The distribution type this strategy will use to generate the noise.

type Sensitivity

The type the sensitivity used for privacy analysis has.

Required Methods

fn from_budget(b: Self::Budget) -> Self

Create a strategy from a differential privacy budget. The distribution created with create_distribution should provide the amount of privacy specified here.

fn create_distribution( &self, s: Self::Sensitivity ) -> Result<Self::Distribution, DpError>

Create a new distribution parametrized s.t. adding samples to the result of a function with sensitivity s will yield differential privacy of the DP variant given in the Budget type. Can error upon invalid parameters.