Module prio::field

Finite field arithmetic.

Basic field arithmetic is captured in the FieldElement trait. Fields used in Prio implement FftFriendlyFieldElement, and have an associated element called the “generator” that generates a multiplicative subgroup of order 2^n for some n.

Structs

Field64

GF(18446744069414584321), a 64-bit field.

Field96Deprecated

GF(79228148845226978974766202881), a 96-bit field.

Field128

GF(340282366920938462946865773367900766209), a 128-bit field.

Field255

GF(2^255 - 19), a 255-bit field.

FieldPrio2

Same as Field32, but encoded in little endian for compatibility with Prio v2.

Enums

FieldError

Possible errors from finite field operations.

Traits

FftFriendlyFieldElement

Objects with this trait represent an element of GF(p), where p is some prime and the field’s multiplicative group has a subgroup with an order that is a power of 2, and at least 2^20.

FieldElement

Objects with this trait represent an element of GF(p) for some prime p.

FieldElementWithInteger

Extension trait for field elements that can be converted back and forth to an integer type.

Functions

random_vector

Generate a vector of uniformly distributed random field elements.