Trait prio::field::FftFriendlyFieldElement

pub trait FftFriendlyFieldElement: FieldElementWithInteger {
    // Required methods
    fn generator_order() -> Self::Integer;
    fn generator() -> Self;
    fn root(l: usize) -> Option<Self>;
}

Objects with this trait represent an element of GF(p), where p is some prime and the field’s multiplicative group has a subgroup with an order that is a power of 2, and at least 2^20.

Required Methods

fn generator_order() -> Self::Integer

Returns the size of the multiplicative subgroup generated by FftFriendlyFieldElement::generator.

fn generator() -> Self

Returns the generator of the multiplicative subgroup of size FftFriendlyFieldElement::generator_order.

fn root(l: usize) -> Option<Self>

Returns the 2^l-th principal root of unity for any l <= 20. Note that the 2^0-th prinicpal root of unity is 1 by definition.