Trait prio::field::FieldElementWithInteger

pub trait FieldElementWithInteger: FieldElement + From<Self::Integer> {
    type IntegerTryFromError: Error;
    type TryIntoU64Error: Error;
    type Integer: Copy + Debug + Eq + Ord + BitAnd<Output = Self::Integer> + Div<Output = Self::Integer> + Shl<usize, Output = Self::Integer> + Shr<usize, Output = Self::Integer> + Add<Output = Self::Integer> + Sub<Output = Self::Integer> + From<Self> + TryFrom<usize, Error = Self::IntegerTryFromError> + TryInto<u64, Error = Self::TryIntoU64Error>;

    // Required methods
    fn pow(&self, exp: Self::Integer) -> Self;
    fn modulus() -> Self::Integer;
}

Extension trait for field elements that can be converted back and forth to an integer type.

The Integer associated type is an integer (primitive or otherwise) that supports various arithmetic operations. The order of the field is guaranteed to fit inside the range of the integer type. This trait also defines methods on field elements, pow and modulus, that make use of the associated integer type.

Required Associated Types

type IntegerTryFromError: Error

The error returned if converting usize to an Integer fails.

type TryIntoU64Error: Error

The error returned if converting an Integer to a u64 fails.

type Integer: Copy + Debug + Eq + Ord + BitAnd<Output = Self::Integer> + Div<Output = Self::Integer> + Shl<usize, Output = Self::Integer> + Shr<usize, Output = Self::Integer> + Add<Output = Self::Integer> + Sub<Output = Self::Integer> + From<Self> + TryFrom<usize, Error = Self::IntegerTryFromError> + TryInto<u64, Error = Self::TryIntoU64Error>

The integer representation of a field element.

Required Methods

fn pow(&self, exp: Self::Integer) -> Self

Modular exponentation, i.e., self^exp (mod p).

fn modulus() -> Self::Integer

Returns the prime modulus p.