Trait prio::flp::Type

pub trait Type: Sized + Eq + Clone + Debug {
    type Measurement: Clone + Debug;
    type AggregateResult: Clone + Debug;
    type Field: FftFriendlyFieldElement;

    const ID: u32;

    // Required methods
    fn encode_measurement(
        &self,
        measurement: &Self::Measurement
    ) -> Result<Vec<Self::Field>, FlpError>;
    fn decode_result(
        &self,
        data: &[Self::Field],
        num_measurements: usize
    ) -> Result<Self::AggregateResult, FlpError>;
    fn gadget(&self) -> Vec<Box<dyn Gadget<Self::Field>>>;
    fn valid(
        &self,
        gadgets: &mut Vec<Box<dyn Gadget<Self::Field>>>,
        input: &[Self::Field],
        joint_rand: &[Self::Field],
        num_shares: usize
    ) -> Result<Self::Field, FlpError>;
    fn truncate(
        &self,
        input: Vec<Self::Field>
    ) -> Result<Vec<Self::Field>, FlpError>;
    fn input_len(&self) -> usize;
    fn proof_len(&self) -> usize;
    fn verifier_len(&self) -> usize;
    fn output_len(&self) -> usize;
    fn joint_rand_len(&self) -> usize;
    fn prove_rand_len(&self) -> usize;
    fn query_rand_len(&self) -> usize;

    // Provided methods
    fn prove(
        &self,
        input: &[Self::Field],
        prove_rand: &[Self::Field],
        joint_rand: &[Self::Field]
    ) -> Result<Vec<Self::Field>, FlpError> { ... }
    fn query(
        &self,
        input: &[Self::Field],
        proof: &[Self::Field],
        query_rand: &[Self::Field],
        joint_rand: &[Self::Field],
        num_shares: usize
    ) -> Result<Vec<Self::Field>, FlpError> { ... }
    fn decide(&self, verifier: &[Self::Field]) -> Result<bool, FlpError> { ... }
    fn valid_call_check(
        &self,
        input: &[Self::Field],
        joint_rand: &[Self::Field]
    ) -> Result<(), FlpError> { ... }
    fn truncate_call_check(&self, input: &[Self::Field]) -> Result<(), FlpError> { ... }

}

A type. Implementations of this trait specify how a particular kind of measurement is encoded as a vector of field elements and how validity of the encoded measurement is determined. Validity is determined via an arithmetic circuit evaluated over the encoded measurement.

Required Associated Types

type Measurement: Clone + Debug

The type of raw measurement to be encoded.

type AggregateResult: Clone + Debug

The type of aggregate result for this type.

type Field: FftFriendlyFieldElement

The finite field used for this type.

Required Associated Constants

const ID: u32

The Prio3 VDAF identifier corresponding to this type.

Required Methods

fn encode_measurement( &self, measurement: &Self::Measurement ) -> Result<Vec<Self::Field>, FlpError>

Encodes a measurement as a vector of Self::input_len field elements.

fn decode_result( &self, data: &[Self::Field], num_measurements: usize ) -> Result<Self::AggregateResult, FlpError>

Decode an aggregate result.

fn gadget(&self) -> Vec<Box<dyn Gadget<Self::Field>>>

Returns the sequence of gadgets associated with the validity circuit.

Notes

The construction of [BBCG+19, Theorem 4.3] uses a single gadget rather than many. The idea to generalize the proof system to allow multiple gadgets is discussed briefly in [BBCG+19, Remark 4.5], but no construction is given. The construction implemented here requires security analysis.

fn valid( &self, gadgets: &mut Vec<Box<dyn Gadget<Self::Field>>>, input: &[Self::Field], joint_rand: &[Self::Field], num_shares: usize ) -> Result<Self::Field, FlpError>

Evaluates the validity circuit on an input and returns the output.

Parameters

gadgets is the sequence of gadgets, presumably output by Self::gadget.

input is the input to be validated.

joint_rand is the joint randomness shared by the prover and verifier.

num_shares is the number of input shares.

Example usage

Applications typically do not call this method directly. It is used internally by Self::prove and Self::query to generate and verify the proof respectively.

use prio::flp::types::Count;
use prio::flp::Type;
use prio::field::{random_vector, FieldElement, Field64};

let count = Count::new();
let input: Vec<Field64> = count.encode_measurement(&1).unwrap();
let joint_rand = random_vector(count.joint_rand_len()).unwrap();
let v = count.valid(&mut count.gadget(), &input, &joint_rand, 1).unwrap();
assert_eq!(v, Field64::zero());

fn truncate( &self, input: Vec<Self::Field> ) -> Result<Vec<Self::Field>, FlpError>

Constructs an aggregatable output from an encoded input. Calling this method is only safe once input has been validated.

fn input_len(&self) -> usize

The length in field elements of the encoded input returned by Self::encode_measurement.

fn proof_len(&self) -> usize

The length in field elements of the proof generated for this type.

fn verifier_len(&self) -> usize

The length in field elements of the verifier message constructed by Self::query.

fn output_len(&self) -> usize

The length of the truncated output (i.e., the output of Type::truncate).

fn joint_rand_len(&self) -> usize

The length of the joint random input.

fn prove_rand_len(&self) -> usize

The length in field elements of the random input consumed by the prover to generate a proof. This is the same as the sum of the arity of each gadget in the validity circuit.

fn query_rand_len(&self) -> usize

The length in field elements of the random input consumed by the verifier to make queries against inputs and proofs. This is the same as the number of gadgets in the validity circuit.

Provided Methods

fn prove( &self, input: &[Self::Field], prove_rand: &[Self::Field], joint_rand: &[Self::Field] ) -> Result<Vec<Self::Field>, FlpError>

Generate a proof of an input’s validity. The return value is a sequence of Self::proof_len field elements.

Parameters

input is the input.

prove_rand is the prover’ randomness.

joint_rand is the randomness shared by the prover and verifier.

fn query( &self, input: &[Self::Field], proof: &[Self::Field], query_rand: &[Self::Field], joint_rand: &[Self::Field], num_shares: usize ) -> Result<Vec<Self::Field>, FlpError>

Query an input and proof and return the verifier message. The return value has length Self::verifier_len.

Parameters

input is the input or input share.

proof is the proof or proof share.

query_rand is the verifier’s randomness.

joint_rand is the randomness shared by the prover and verifier.

num_shares is the total number of input shares.

fn decide(&self, verifier: &[Self::Field]) -> Result<bool, FlpError>

Returns true if the verifier message indicates that the input from which it was generated is valid.

fn valid_call_check( &self, input: &[Self::Field], joint_rand: &[Self::Field] ) -> Result<(), FlpError>

Check whether input and joint_rand have the length expected by self, return FlpError::Valid otherwise.

fn truncate_call_check(&self, input: &[Self::Field]) -> Result<(), FlpError>

Check if the length of input matches self’s input_len(), return FlpError::Truncate otherwise.