Trait prio::flp::TypeWithNoise

pub trait TypeWithNoise<S>: Typewhere
    S: DifferentialPrivacyStrategy,{
    // Required method
    fn add_noise_to_result(
        &self,
        dp_strategy: &S,
        agg_result: &mut [Self::Field],
        num_measurements: usize
    ) -> Result<(), FlpError>;
}

A type which supports adding noise to aggregate shares for Server Differential Privacy.

Required Methods

fn add_noise_to_result( &self, dp_strategy: &S, agg_result: &mut [Self::Field], num_measurements: usize ) -> Result<(), FlpError>

Add noise to the aggregate share to obtain differential privacy.