Module prio::flp::types::fixedpoint_l2

A Type for summing vectors of fixed point numbers where the L2 norm of each vector is bounded by 1 and adding discrete Gaussian noise in order to achieve server differential privacy.

In the following a high level overview over the inner workings of this type is given and implementation details are discussed. It is not necessary for using the type, but it should be helpful when trying to understand the implementation.

Overview

Clients submit a vector of numbers whose values semantically lie in [-1,1), together with a norm in the range [0,1). The validation circuit checks that the norm of the vector is equal to the submitted norm, while the encoding guarantees that the submitted norm lies in the correct range.

The bound on the L2 norm allows calibration of discrete Gaussian noise added after aggregation, making the procedure differentially private.

Submission layout

The client submissions contain a share of their vector and the norm they claim it has. The submission is a vector of field elements laid out as follows:

|---- bits_per_entry * entries ----|---- bits_for_norm ----|
 ^                                  ^
 \- the input vector entries        |
                                    \- the encoded norm

Different number encodings

Let n denote the number of bits of the chosen fixed-point type. Numbers occur in 5 different representations:

Clients have a vector whose entries are fixed point numbers. Only those fixed point types are supported where the numbers lie in the range [-1,1).

Because norm computation happens in the validation circuit, it is done on entries encoded as field elements. That is, the same vector entries are now represented by integers in the range [0,2^n), where -1 is represented by 0 and +1 by 2^n.

Because the field is not necessarily exactly of size 2^n, but might be larger, it is not enough to encode a vector entry as in (2.) and submit it to the aggregator. Instead, in order to make sure that all submitted values are in the correct range, they are bit-encoded. (This is the same as what happens in the Sum type.) This means that instead of sending a field element in the range [0,2^n), we send n field elements representing the bit encoding. The validation circuit can verify that all submitted “bits” are indeed either 0 or 1.

The computed and submitted norms are treated similar to the vector entries, but they have a different number of bits, namely 2n-2.

As the aggregation result is a pointwise sum of the client vectors, the numbers no longer (semantically) lie in the range [-1,1), and cannot be represented by the same fixed point type as the input. Instead the decoding happens directly into a vector of floats.

Fixed point encoding

Submissions consist of encoded fixed-point numbers in [-1,1) represented as field elements in [0,2^n), where n is the number of bits the fixed-point representation has. Encoding and decoding is handled by the associated functions of the CompatibleFloat trait. Semantically, the following function describes how a fixed-point value x in range [-1,1) is converted to a field integer:

enc : [-1,1) -> [0,2^n)
enc(x) = 2^(n-1) * x + 2^(n-1)

The inverse is:

dec : [0,2^n) -> [-1,1)
dec(y) = (y - 2^(n-1)) * 2^(1-n)

Note that these functions only make sense when interpreting all occuring numbers as real numbers. Since our signed fixed-point numbers are encoded as two’s complement integers, the computation that happens in CompatibleFloat::to_field_integer is actually simpler.

Value 1

We actually do not allow the submitted norm or vector entries to be exactly 1, but rather require them to be strictly less. Supporting 1 would entail a more fiddly encoding and is not necessary for our usecase. The largest representable vector entry can be computed by dec(2^n-1). For example, it is 0.999969482421875 for n = 16.

Norm computation

The L2 norm of a vector xs of numbers in [-1,1) is given by:

norm(xs) = sqrt(sum_{x in xs} x^2)

Instead of computing the norm, we make two simplifications:

We ignore the square root, which means that we are actually computing the square of the norm.

We want our norm computation result to be integral and in the range [0, 2^(2n-2)), so we can represent it in our field integers. We achieve this by multiplying with 2^(2n-2). This means that what is actually computed in this type is the following:

our_norm(xs) = 2^(2n-2) * norm(xs)^2

Explained more visually, our_norm() is a composition of three functions:

                   map of dec()                    norm()          "mult with 2^(2n-2)"
  vector of [0,2^n)    ->       vector of [-1,1)    ->       [0,1)         ->           [0,2^(2n-2))
                                        ^                      ^
                                        |                      |
                    fractions with denom of 2^(n-1)     fractions with denom of 2^(2n-2)

(Note that the ranges on the LHS and RHS of "mult with 2^(2n-2)" are stated here for vectors with a norm less than 1.)

Given a vector ys of numbers in the field integer encoding (in [0,2^n)), this gives the following equation:

our_norm_on_encoded(ys) = our_norm([dec(y) for y in ys])
                        = 2^(2n-2) * sum_{y in ys} ((y - 2^(n-1)) * 2^(1-n))^2
                        = 2^(2n-2)
                          * sum_{y in ys} y^2 - 2*y*2^(n-1) + (2^(n-1))^2
                          * 2^(1-n)^2
                        = sum_{y in ys} y^2 - (2^n)*y + 2^(2n-2)

Let d denote the number of the vector entries. The maximal value the result of our_norm_on_encoded() can take occurs in the case where all entries are 2^n-1, in which case d * 2^(2n-2) is an upper bound to the result. The finite field used for encoding must be at least as large as this. For validating that the norm of the submitted vector lies in the correct range, consider the following:

The result of norm(xs) should be in [0,1).

Thus, the result of our_norm(xs) should be in [0,2^(2n-2)).

The result of our_norm_on_encoded(ys) should be in [0,2^(2n-2)). This means that the valid norms are exactly those representable with 2n-2 bits.

Noise and Differential Privacy

Bounding the submission norm bounds the impact that changing a single client’s submission can have on the aggregate. That is, the so-called L2-sensitivity of the procedure is equal to two times the norm bound, namely 2^n. Therefore, adding discrete Gaussian noise with standard deviation sigma = (2^n)/epsilonfor someepsilon will make the procedure [(epsilon^2)/2zero-concentrated differentially private](https://arxiv.org/abs/2004.00010).epsilonis given as a parameter to theadd_noise_to_resultfunction, as part of thedp_strategy argument of type [ZCdpDiscreteGaussian`].

Differences in the computation because of distribution

In decode_result(), what is decoded are not the submitted entries of a single client, but the sum of the the entries of all clients. We have to take this into account, and cannot directly use the dec() function from above. Instead we use:

dec'(y) = y * 2^(1-n) - c

Here, c is the number of clients.

Naming in the implementation

The following names are used:

self.bits_per_entry is n

self.entries is d

self.bits_for_norm is 2n-2

Modules

compatible_float

Implementations of encoding fixed point types as field elements and field elements as floats for the FixedPointBoundedL2VecSum type.

Structs

FixedPointBoundedL2VecSum

The fixed point vector sum data type. Each measurement is a vector of fixed point numbers of type T, and the aggregate result is the float vector of the sum of the measurements.