Module prio::idpf

This module implements the incremental distributed point function (IDPF) described in [draft-irtf-cfrg-vdaf-07].

Structs

HashMapCache

A simple IdpfCache implementation that caches intermediate results in an in-memory hash map, with no eviction.

Idpf

This defines a family of IDPFs (incremental distributed point functions) with certain types of values at inner tree nodes and at leaf tree nodes.

IdpfInput

An index used as the input to an IDPF evaluation.

IdpfPublicShare

An IDPF public share. This contains the list of correction words used by all parties when evaluating the IDPF.

NoCache

A no-op IdpfCache implementation that always reports a cache miss.

RingBufferCache

A simple IdpfCache implementation that caches intermediate results in memory, with first-in-first-out eviction, and lookups via linear probing.

Enums

IdpfError

IDPF-related errors.

IdpfOutputShare

An output from evaluation of an IDPF at some level and index.

Traits

IdpfCache

An interface that provides memoization of IDPF computations.

IdpfValue

Trait for values to be programmed into an IDPF.