Trait prio::topology::ping_pong::PingPongTopology

pub trait PingPongTopology<const VERIFY_KEY_SIZE: usize, const NONCE_SIZE: usize>: Aggregator<VERIFY_KEY_SIZE, NONCE_SIZE> {
    type State;
    type ContinuedValue;
    type Transition;

    // Required methods
    fn leader_initialized(
        &self,
        verify_key: &[u8; VERIFY_KEY_SIZE],
        agg_param: &Self::AggregationParam,
        nonce: &[u8; NONCE_SIZE],
        public_share: &Self::PublicShare,
        input_share: &Self::InputShare
    ) -> Result<(Self::State, PingPongMessage), PingPongError>;
    fn helper_initialized(
        &self,
        verify_key: &[u8; VERIFY_KEY_SIZE],
        agg_param: &Self::AggregationParam,
        nonce: &[u8; NONCE_SIZE],
        public_share: &Self::PublicShare,
        input_share: &Self::InputShare,
        inbound: &PingPongMessage
    ) -> Result<PingPongTransition<VERIFY_KEY_SIZE, NONCE_SIZE, Self>, PingPongError>;
    fn leader_continued(
        &self,
        leader_state: Self::State,
        agg_param: &Self::AggregationParam,
        inbound: &PingPongMessage
    ) -> Result<Self::ContinuedValue, PingPongError>;
    fn helper_continued(
        &self,
        helper_state: Self::State,
        agg_param: &Self::AggregationParam,
        inbound: &PingPongMessage
    ) -> Result<Self::ContinuedValue, PingPongError>;
}

Extension trait on crate::vdaf::Aggregator which adds the VDAF Ping-Pong Topology.

Required Associated Types

type State

Specialization of PingPongState for this VDAF.

type ContinuedValue

Specialization of PingPongContinuedValue for this VDAF.

type Transition

Specializaton of PingPongTransition for this VDAF.

Required Methods

fn leader_initialized( &self, verify_key: &[u8; VERIFY_KEY_SIZE], agg_param: &Self::AggregationParam, nonce: &[u8; NONCE_SIZE], public_share: &Self::PublicShare, input_share: &Self::InputShare ) -> Result<(Self::State, PingPongMessage), PingPongError>

Initialize leader state using the leader’s input share. Corresponds to ping_pong_leader_init in VDAF.

If successful, the returned PingPongMessage (which will always be PingPongMessage::Initialize) should be transmitted to the helper. The returned PingPongState (which will always be PingPongState::Continued) should be used by the leader along with the next PingPongMessage received from the helper as input to Self::leader_continued to advance to the next round.

fn helper_initialized( &self, verify_key: &[u8; VERIFY_KEY_SIZE], agg_param: &Self::AggregationParam, nonce: &[u8; NONCE_SIZE], public_share: &Self::PublicShare, input_share: &Self::InputShare, inbound: &PingPongMessage ) -> Result<PingPongTransition<VERIFY_KEY_SIZE, NONCE_SIZE, Self>, PingPongError>

Initialize helper state using the helper’s input share and the leader’s first prepare share. Corresponds to ping_pong_helper_init in the forthcoming draft-irtf-cfrg-vdaf-07.

If successful, the returned PingPongTransition should be evaluated, yielding a PingPongMessage, which should be transmitted to the leader, and a PingPongState.

If the state is PingPongState::Continued, then it should be used by the helper along with the next PingPongMessage received from the leader as input to Self::helper_continued to advance to the next round. The helper may store the PingPongTransition between rounds of preparation instead of the PingPongState and PingPongMessage.

If the state is PingPongState::Finished, then preparation is finished and the output share may be accumulated.

Errors

inbound must be PingPongMessage::Initialize or the function will fail.

fn leader_continued( &self, leader_state: Self::State, agg_param: &Self::AggregationParam, inbound: &PingPongMessage ) -> Result<Self::ContinuedValue, PingPongError>

Continue preparation based on the leader’s current state and an incoming PingPongMessage from the helper. Corresponds to ping_pong_leader_continued in VDAF.

If successful, the returned PingPongContinuedValue will either be:

PingPongContinuedValue::WithMessage { transition }: transition should be evaluated, yielding a PingPongMessage, which should be transmitted to the helper, and a PingPongState.

If the state is PingPongState::Continued, then it should be used by the leader along with the next PingPongMessage received from the helper as input to Self::leader_continued to advance to the next round. The leader may store the PingPongTransition between rounds of preparation instead of of the PingPongState and PingPongMessage.

If the state is PingPongState::Finished, then preparation is finished and the output share may be accumulated.

PingPongContinuedValue::FinishedNoMessage: preparation is finished and the output share may be accumulated. No message needs to be sent to the helper.

Errors

leader_state must be PingPongState::Continued or the function will fail.

inbound must not be PingPongMessage::Initialize or the function will fail.

Notes

The specification of this function in VDAF takes the aggregation parameter. This version does not, because crate::vdaf::Aggregator::prepare_preprocess does not take the aggregation parameter. This may change in the future if/when #670 is addressed.

fn helper_continued( &self, helper_state: Self::State, agg_param: &Self::AggregationParam, inbound: &PingPongMessage ) -> Result<Self::ContinuedValue, PingPongError>

PingPongContinue preparation based on the helper’s current state and an incoming PingPongMessage from the leader. Corresponds to ping_pong_helper_contnued in VDAF.

If successful, the returned PingPongContinuedValue will either be:

PingPongContinuedValue::WithMessage { transition }: transition should be evaluated, yielding a PingPongMessage, which should be transmitted to the leader, and a PingPongState.

If the state is PingPongState::Continued, then it should be used by the helper along with the next PingPongMessage received from the leader as input to Self::helper_continued to advance to the next round. The helper may store the PingPongTransition between rounds of preparation instead of the PingPongState and PingPongMessage.

If the state is PingPongState::Finished, then preparation is finished and the output share may be accumulated.

PingPongContinuedValue::FinishedNoMessage: preparation is finished and the output share may be accumulated. No message needs to be sent to the leader.

Errors

helper_state must be PingPongState::Continued or the function will fail.

inbound must not be PingPongMessage::Initialize or the function will fail.

Notes

The specification of this function in VDAF takes the aggregation parameter. This version does not, because crate::vdaf::Aggregator::prepare_preprocess does not take the aggregation parameter. This may change in the future if/when #670 is addressed.