Module prio::vdaf

Verifiable Distributed Aggregation Functions (VDAFs) as described in [draft-irtf-cfrg-vdaf-07].

Modules

poplar1

Implementation of Poplar1 as specified in [draft-irtf-cfrg-vdaf-07].

prio3

Implementation of the Prio3 VDAF [draft-irtf-cfrg-vdaf-07].

xof

Implementations of XOFs specified in [draft-irtf-cfrg-vdaf-07].

Structs

AggregateShare

An aggregate share comprised of a vector of field elements.

OutputShare

An output share comprised of a vector of field elements.

Enums

PrepareTransition

A state transition of an Aggregator during the Prepare process.

Share

An additive share of a vector of field elements.

VdafError

Errors emitted by this module.

Traits

Aggregatable

An aggregate share resulting from aggregating output shares together that can merged with aggregate shares of the same type.

Aggregator

The Aggregator’s role in the execution of a VDAF.

AggregatorWithNoise

Aggregator that implements differential privacy with Aggregator-side noise addition.

Client

The Client’s role in the execution of a VDAF.

Collector

The Collector’s role in the execution of a VDAF.

Vdaf

The base trait for VDAF schemes. This trait is inherited by traits Client, Aggregator, and Collector, which define the roles of the various parties involved in the execution of the VDAF.