Struct prio::vdaf::poplar1::Poplar1

pub struct Poplar1<P, const SEED_SIZE: usize> { /* private fields */ }

The Poplar1 VDAF.

Implementations

impl<P, const SEED_SIZE: usize> Poplar1<P, SEED_SIZE>

pub fn new(bits: usize) -> Self

Create an instance of Poplar1. The caller provides the bit length of each measurement (BITS as defined in the [draft-irtf-cfrg-vdaf-07]).

impl Poplar1<XofShake128, 16>

pub fn new_shake128(bits: usize) -> Self

Create an instance of Poplar1 using XofShake128. The caller provides the bit length of each measurement (BITS as defined in the [draft-irtf-cfrg-vdaf-07]).