Module prio::vdaf::prg

Implementations of PRGs specified in [draft-irtf-cfrg-vdaf-05].

Structs

PrgAes128Deprecated

The PRG based on AES128 as specified in previous versions of draft-irtf-cfrg-vdaf.

PrgFixedKeyAes128

PrgFixedKeyAes128 as specified in [draft-irtf-cfrg-vdaf-05]. This PRG is NOT RECOMMENDED for general use; see Section 9 (“Security Considerations”) for details.

PrgFixedKeyAes128Key

Factory to produce multiple PrgFixedKeyAes128 instances with the same fixed key and different seeds.

PrgSha3

The PRG based on SHA-3 as specified in [draft-irtf-cfrg-vdaf-05].

Seed

Input of Prg.

SeedStreamAes128

The key stream produced by AES128 in CTR-mode.

SeedStreamFixedKeyAes128

Seed stream for PrgFixedKeyAes128.

SeedStreamSha3

The key stream produced by the cSHAKE128 XOF.

Traits

CoinToss

Types implementing CoinToss can be randomly sampled from a SeedStream.

Prg

A pseudorandom generator (PRG) with the interface specified in [draft-irtf-cfrg-vdaf-05].

SeedStream

A stream of pseudorandom bytes derived from a seed.