Trait prio::vdaf::prg::Prg

pub trait Prg<const SEED_SIZE: usize>: Clone + Debug {
    type SeedStream: SeedStream;

    fn init(seed_bytes: &[u8; SEED_SIZE], custom: &[u8]) -> Self;
    fn update(&mut self, data: &[u8]);
    fn into_seed_stream(self) -> Self::SeedStream;

    fn into_seed(self) -> Seed<SEED_SIZE> { ... }
    fn seed_stream(
        seed: &Seed<SEED_SIZE>,
        custom: &[u8],
        binder: &[u8]
    ) -> Self::SeedStream { ... }
}

A pseudorandom generator (PRG) with the interface specified in [draft-irtf-cfrg-vdaf-05].

Required Associated Types

type SeedStream: SeedStream

The type of stream produced by this PRG.

Required Methods

fn init(seed_bytes: &[u8; SEED_SIZE], custom: &[u8]) -> Self

Construct an instance of Prg with the given seed.

fn update(&mut self, data: &[u8])

Update the PRG state by passing in the next fragment of the info string. The final info string is assembled from the concatenation of sequence of fragments passed to this method.

fn into_seed_stream(self) -> Self::SeedStream

Finalize the PRG state, producing a seed stream.

Provided Methods

fn into_seed(self) -> Seed<SEED_SIZE>

Finalize the PRG state, producing a seed.

fn seed_stream(
seed: &Seed<SEED_SIZE>,
custom: &[u8],
binder: &[u8]
) -> Self::SeedStream

Construct a seed stream from the given seed and info string.