Type Definition prio::vdaf::prio3::Prio3Histogram

pub type Prio3Histogram = Prio3<Histogram<Field128, ParallelSum<Field128, Mul<Field128>>>, XofShake128, 16>;

The histogram type. Each measurement is an integer in [0, length) and the result is a histogram counting the number of occurrences of each measurement.

Implementations

impl Prio3Histogram

pub fn new_histogram( num_aggregators: u8, length: usize, chunk_length: usize ) -> Result<Self, VdafError>

Constructs an instance of Prio3Histogram with the given number of aggregators, number of buckets, and parallel sum gadget chunk length.