Trait prio::vdaf::Aggregatable

pub trait Aggregatable: Clone + Debug + From<Self::OutputShare> {
    type OutputShare;

    // Required methods
    fn merge(&mut self, agg_share: &Self) -> Result<(), VdafError>;
    fn accumulate(
        &mut self,
        output_share: &Self::OutputShare
    ) -> Result<(), VdafError>;
}

An aggregate share resulting from aggregating output shares together that can merged with aggregate shares of the same type.

Required Associated Types

type OutputShare

Type of output shares that can be accumulated into an aggregate share.

Required Methods

fn merge(&mut self, agg_share: &Self) -> Result<(), VdafError>

Update an aggregate share by merging it with another (agg_share).

fn accumulate( &mut self, output_share: &Self::OutputShare ) -> Result<(), VdafError>

Update an aggregate share by adding output_share.