Trait prio::vdaf::Aggregator

pub trait Aggregator<const VERIFY_KEY_SIZE: usize, const NONCE_SIZE: usize>: Vdaf {
    type PrepareState: Clone + Debug + PartialEq + Eq;
    type PrepareShare: Clone + Debug + ParameterizedDecode<Self::PrepareState> + Encode;
    type PrepareMessage: Clone + Debug + PartialEq + Eq + ParameterizedDecode<Self::PrepareState> + Encode;

    // Required methods
    fn prepare_init(
        &self,
        verify_key: &[u8; VERIFY_KEY_SIZE],
        agg_id: usize,
        agg_param: &Self::AggregationParam,
        nonce: &[u8; NONCE_SIZE],
        public_share: &Self::PublicShare,
        input_share: &Self::InputShare
    ) -> Result<(Self::PrepareState, Self::PrepareShare), VdafError>;
    fn prepare_shares_to_prepare_message<M: IntoIterator<Item = Self::PrepareShare>>(
        &self,
        agg_param: &Self::AggregationParam,
        inputs: M
    ) -> Result<Self::PrepareMessage, VdafError>;
    fn prepare_next(
        &self,
        state: Self::PrepareState,
        input: Self::PrepareMessage
    ) -> Result<PrepareTransition<Self, VERIFY_KEY_SIZE, NONCE_SIZE>, VdafError>;
    fn aggregate<M: IntoIterator<Item = Self::OutputShare>>(
        &self,
        agg_param: &Self::AggregationParam,
        output_shares: M
    ) -> Result<Self::AggregateShare, VdafError>;

    // Provided methods
    fn prepare_preprocess<M: IntoIterator<Item = Self::PrepareShare>>(
        &self,
        agg_param: &Self::AggregationParam,
        inputs: M
    ) -> Result<Self::PrepareMessage, VdafError> { ... }
    fn prepare_step(
        &self,
        state: Self::PrepareState,
        input: Self::PrepareMessage
    ) -> Result<PrepareTransition<Self, VERIFY_KEY_SIZE, NONCE_SIZE>, VdafError> { ... }
}

The Aggregator’s role in the execution of a VDAF.

Required Associated Types

type PrepareState: Clone + Debug + PartialEq + Eq

State of the Aggregator during the Prepare process.

type PrepareShare: Clone + Debug + ParameterizedDecode<Self::PrepareState> + Encode

The type of messages sent by each aggregator at each round of the Prepare Process.

Decoding takes a Self::PrepareState as a parameter; this Self::PrepareState may be associated with any aggregator involved in the execution of the VDAF.

type PrepareMessage: Clone + Debug + PartialEq + Eq + ParameterizedDecode<Self::PrepareState> + Encode

Result of preprocessing a round of preparation shares. This is used by all aggregators as an input to the next round of the Prepare Process.

Decoding takes a Self::PrepareState as a parameter; this Self::PrepareState may be associated with any aggregator involved in the execution of the VDAF.

Required Methods

fn prepare_init( &self, verify_key: &[u8; VERIFY_KEY_SIZE], agg_id: usize, agg_param: &Self::AggregationParam, nonce: &[u8; NONCE_SIZE], public_share: &Self::PublicShare, input_share: &Self::InputShare ) -> Result<(Self::PrepareState, Self::PrepareShare), VdafError>

Begins the Prepare process with the other Aggregators. The Self::PrepareState returned is passed to Self::prepare_next to get this aggregator’s first-round prepare message.

Implements Vdaf.prep_init from VDAF.

fn prepare_shares_to_prepare_message<M: IntoIterator<Item = Self::PrepareShare>>( &self, agg_param: &Self::AggregationParam, inputs: M ) -> Result<Self::PrepareMessage, VdafError>

Preprocess a round of preparation shares into a single input to Self::prepare_next.

Implements Vdaf.prep_shares_to_prep from VDAF.

fn prepare_next( &self, state: Self::PrepareState, input: Self::PrepareMessage ) -> Result<PrepareTransition<Self, VERIFY_KEY_SIZE, NONCE_SIZE>, VdafError>

Compute the next state transition from the current state and the previous round of input messages. If this returns PrepareTransition::Continue, then the returned Self::PrepareShare should be combined with the other Aggregators’ PrepareShares from this round and passed into another call to this method. This continues until this method returns PrepareTransition::Finish, at which point the returned output share may be aggregated. If the method returns an error, the aggregator should consider its input share invalid and not attempt to process it any further.

Implements Vdaf.prep_next from VDAF.

fn aggregate<M: IntoIterator<Item = Self::OutputShare>>( &self, agg_param: &Self::AggregationParam, output_shares: M ) -> Result<Self::AggregateShare, VdafError>

Aggregates a sequence of output shares into an aggregate share.

Provided Methods

fn prepare_preprocess<M: IntoIterator<Item = Self::PrepareShare>>( &self, agg_param: &Self::AggregationParam, inputs: M ) -> Result<Self::PrepareMessage, VdafError>

👎Deprecated since 0.15.0: Use Vdaf::prepare_shares_to_prepare_message instead

Preprocess a round of preparation shares into a single input to Self::prepare_next.

Implements Vdaf.prep_shares_to_prep from VDAF.

Notes

Self::prepare_shares_to_prepare_message is preferable since its name better matches the specification.

fn prepare_step( &self, state: Self::PrepareState, input: Self::PrepareMessage ) -> Result<PrepareTransition<Self, VERIFY_KEY_SIZE, NONCE_SIZE>, VdafError>

👎Deprecated since 0.15.0: Use Vdaf::prepare_next

Compute the next state transition from the current state and the previous round of input messages. If this returns PrepareTransition::Continue, then the returned Self::PrepareShare should be combined with the other Aggregators’ PrepareShares from this round and passed into another call to this method. This continues until this method returns PrepareTransition::Finish, at which point the returned output share may be aggregated. If the method returns an error, the aggregator should consider its input share invalid and not attempt to process it any further.

Implements Vdaf.prep_next from VDAF.

Notes

Self::prepare_next is preferable since its name better matches the specification.