Trait prio::vdaf::Vdaf

pub trait Vdaf: Clone + Debug {
    type Measurement: Clone + Debug;
    type AggregateResult: Clone + Debug;
    type AggregationParam: Clone + Debug + Decode + Encode;
    type PublicShare: Clone + Debug + ParameterizedDecode<Self> + Encode;
    type InputShare: Clone + Debug + for<'a> ParameterizedDecode<(&'a Self, usize)> + Encode;
    type OutputShare: Clone + Debug + for<'a> ParameterizedDecode<(&'a Self, &'a Self::AggregationParam)> + Encode;
    type AggregateShare: Aggregatable<OutputShare = Self::OutputShare> + for<'a> ParameterizedDecode<(&'a Self, &'a Self::AggregationParam)> + Encode;

    const ID: u32;

    // Required method
    fn num_aggregators(&self) -> usize;

    // Provided method
    fn domain_separation_tag(usage: u16) -> [u8; 8] { ... }
}

The base trait for VDAF schemes. This trait is inherited by traits Client, Aggregator, and Collector, which define the roles of the various parties involved in the execution of the VDAF.

Required Associated Types

type Measurement: Clone + Debug

The type of Client measurement to be aggregated.

type AggregateResult: Clone + Debug

The aggregate result of the VDAF execution.

type AggregationParam: Clone + Debug + Decode + Encode

The aggregation parameter, used by the Aggregators to map their input shares to output shares.

type PublicShare: Clone + Debug + ParameterizedDecode<Self> + Encode

A public share sent by a Client.

type InputShare: Clone + Debug + for<'a> ParameterizedDecode<(&'a Self, usize)> + Encode

An input share sent by a Client.

type OutputShare: Clone + Debug + for<'a> ParameterizedDecode<(&'a Self, &'a Self::AggregationParam)> + Encode

An output share recovered from an input share by an Aggregator.

type AggregateShare: Aggregatable<OutputShare = Self::OutputShare> + for<'a> ParameterizedDecode<(&'a Self, &'a Self::AggregationParam)> + Encode

An Aggregator’s share of the aggregate result.

Required Associated Constants

const ID: u32

Algorithm identifier for this VDAF.

Required Methods

fn num_aggregators(&self) -> usize

The number of Aggregators. The Client generates as many input shares as there are Aggregators.

Provided Methods

fn domain_separation_tag(usage: u16) -> [u8; 8]

Generate the domain separation tag for this VDAF. The output is used for domain separation by the XOF.