Trait prio::codec::Encode

pub trait Encode {
    fn encode(&self, bytes: &mut Vec<u8>);

    fn get_encoded(&self) -> Vec<u8> { ... }
    fn encoded_len(&self) -> Option<usize> { ... }
}

Describes how to encode objects into a byte sequence.

Required Methods

fn encode(&self, bytes: &mut Vec<u8>)

Append the encoded form of this object to the end of bytes, growing the vector as needed.

Provided Methods

fn get_encoded(&self) -> Vec<u8>

Convenience method to encode a value into a new Vec<u8>.

fn encoded_len(&self) -> Option<usize>

Returns an optional hint indicating how many bytes will be required to encode this value, or None by default.