Trait prio::field::FieldElement

pub trait FieldElement: Sized + Debug + Copy + PartialEq + Eq + ConstantTimeEq + ConditionallySelectable + ConditionallyNegatable + Add<Output = Self> + AddAssign + Sub<Output = Self> + SubAssign + Mul<Output = Self> + MulAssign + Div<Output = Self> + DivAssign + Neg<Output = Self> + Display + for<'a> TryFrom<&'a [u8], Error = FieldError> + Into<Vec<u8>> + Serialize + DeserializeOwned + Encode + Decode + 'static {
    const ENCODED_SIZE: usize;

    fn inv(&self) -> Self;
    fn zero() -> Self;
    fn one() -> Self;

    fn slice_into_byte_vec(values: &[Self]) -> Vec<u8> { ... }
    fn byte_slice_into_vec(bytes: &[u8]) -> Result<Vec<Self>, FieldError> { ... }
}

Objects with this trait represent an element of GF(p) for some prime p.

Required Associated Constants

const ENCODED_SIZE: usize

Size in bytes of an encoded field element.

Required Methods

fn inv(&self) -> Self

Modular inversion, i.e., self^-1 (mod p). If self is 0, then the output is undefined.

fn zero() -> Self

Returns the additive identity.

fn one() -> Self

Returns the multiplicative identity.

Provided Methods

fn slice_into_byte_vec(values: &[Self]) -> Vec<u8>

Convert a slice of field elements into a vector of bytes.

Notes

Ideally we would implement From<&[F: FieldElement]> for Vec<u8> or the corresponding Into, but the orphan rule and the stdlib’s blanket implementations of Into make this impossible.

fn byte_slice_into_vec(bytes: &[u8]) -> Result<Vec<Self>, FieldError>

Convert a slice of bytes into a vector of field elements. The slice is interpreted as a sequence of Self::ENCODED_SIZE-byte sequences.

Errors

Returns an error if the length of the provided byte slice is not a multiple of the size of a field element, or if any of the values in the byte slice are invalid encodings of a field element, because the encoded integer is larger than or equal to the field modulus.

Notes

Ideally we would implement From<&[u8]> for Vec<F: FieldElement> or the corresponding Into, but the orphan rule and the stdlib’s blanket implementations of Into make this impossible.