Module prio::flp::types

A collection of Type implementations.

Modules

fixedpoint_l2

A Type for summing vectors of fixed point numbers where the L2 norm of each vector is bounded by 1.

Structs

Average

The average type. Each measurement is an integer in [0,2^bits) for some 0 < bits < 64 and the aggregate is the arithmetic average.

Count

The counter data type. Each measurement is 0 or 1 and the aggregate result is the sum of the measurements (i.e., the total number of 1s).

Histogram

The histogram type. Each measurement is a non-negative integer and the aggregate is a histogram approximating the distribution of the measurements.

Sum

This sum type. Each measurement is a integer in [0, 2^bits) and the aggregate is the sum of the measurements.

SumVec

A sequence of integers in range [0, 2^bits). This type uses a neat trick from [BBCG+19, Corollary 4.9] to reduce the proof size to roughly the square root of the input size.