Trait prio::vdaf::Client

pub trait Client<const NONCE_SIZE: usize>: Vdaf {
    // Required method
    fn shard(
        &self,
        measurement: &Self::Measurement,
        nonce: &[u8; NONCE_SIZE]
    ) -> Result<(Self::PublicShare, Vec<Self::InputShare>), VdafError>;
}

The Client’s role in the execution of a VDAF.

Required Methods

fn shard( &self, measurement: &Self::Measurement, nonce: &[u8; NONCE_SIZE] ) -> Result<(Self::PublicShare, Vec<Self::InputShare>), VdafError>

Shards a measurement into a public share and a sequence of input shares, one for each Aggregator.

Implements Vdaf::shard from VDAF.